pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-trader"
version = "0.1.0"
description = "Lean cloud-native cryptocurrency trading service for Aster DEX"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.30.4",
    "aiohttp>=3.11.10",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0,<2.12",
    "google-cloud-secret-manager>=2.25.0,<2.26",
    "google-cloud-pubsub>=2.19.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-aiplatform>=1.60.0",
    "vertexai>=1.60.0",
    "redis>=5.1.1",
    "httpx>=0.28.1",
    "orjson>=3.11.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "structlog>=25.5.0",
    "python-telegram-bot>=21.0,<22.6",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pandas-ta-openbb>=0.4.22",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "accelerate>=0.28.0",
    "matplotlib>=3.8.0",
    "google-generativeai>=0.4.0",
    "hyperliquid-python-sdk>=0.21.0",
    "eth-account>=0.11.0",
    "solana>=0.34.0",
    "solders>=0.18.0",
    "base58>=2.1.0",
]

[project.optional-dependencies]
analytics = [
    "feast>=0.44.1",
    "sentence-transformers>=2.6.1",
    "optuna>=3.6.1",
    "panel>=1.4.4",
    "plotly>=5.22.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/rari-trade"
Repository = "https://github.com/yourusername/rari-trade"
Documentation = "https://rari-trade.readthedocs.io/"
Issues = "https://github.com/yourusername/rari-trade/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cloud_trader*"]



[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | rari_trade_env
  | rari_trade_env_fresh
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["cloud_trader"]
skip = ["__pycache__"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "rari_trade_env",
    "rari_trade_env_fresh",
    "*.egg-info",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "google.cloud.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B601"]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q"
testpaths = ["tests"]
python_files = ["test_*.py"]

asyncio_mode = "auto"
markers = ["asyncio: mark test as asyncio-aware"]
[tool.coverage.run]
source = ["cloud_trader"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
